+++ stub.cpp
@@ -124,7 +124,7 @@ Uint16 actor_to_ui16(const objtype* actor)
     ptrdiff_t index = actor - objlist;
 
     if (index >= 0) {
-        int offset = OBJLIST_OFFSET + (index * OLD_OBJTYPE_SIZE);
+        ptrdiff_t offset = OBJLIST_OFFSET + (index * OLD_OBJTYPE_SIZE);
         assert(offset <= 0xFFFF);
         return static_cast<Uint16>(offset);
     }
@@ -150,7 +150,7 @@ Uint16 static_object_to_ui16(const statobj_t* static_object)
     ptrdiff_t index = static_object - statobjlist;
 
     if (index >= 0) {
-        int offset = STATOBJLIST_OFFSET + (index * OLD_STATOBJ_T_SIZE);
+        ptrdiff_t offset = STATOBJLIST_OFFSET + (index * OLD_STATOBJ_T_SIZE);
         assert(offset <= 0xFFFF);
         return static_cast<Uint16>(offset);
     }
@@ -176,7 +176,7 @@ Uint16 door_object_to_ui16(const doorobj_t* door_object)
     ptrdiff_t index = door_object - doorobjlist;
 
     if (index >= 0) {
-        int offset = DOOROBJLIST_OFFSET + (index * OLD_DOOROBJ_T_SIZE);
+        ptrdiff_t offset = DOOROBJLIST_OFFSET + (index * OLD_DOOROBJ_T_SIZE);
         assert(offset <= 0xFFFF);
         return static_cast<Uint16>(offset);
     }
